refactor rosinput
