+++ lib/rosinput/pcl/pcl.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<description>
+        <libname>pcl</libname>
+        <functions>
+                <function>
+                        <name>Lidar2D</name>
+                        <output type="MATRIX">
+                        </output>
+                        <inputs>
+                                <input  type="STRING">
+                                        <name>topic_name</name>
+                                        <desc>Name of the topic to which the function subscribes </desc>
+                                </input>
+                                <input  type="SCALAR_SCALAR">
+                                        <name>size_queue</name>
+                                        <desc>Size of the queue</desc>
+                                </input>
+                                <input  type="SCALAR_SCALAR">
+                                        <name>sleep</name>
+                                        <desc>sleep duration [-1 : block until receive data ; 0 non-blocking call; N > 0 : sleep during N milliseconds]</desc>
+                                </input>
+                                <input  type="SCALAR_SCALAR">
+                                        <name>range_max</name>
+                                        <desc>Max Range [in meters] : values are projected between 0 and range_max</desc>
+                                </input>
+                        </inputs>
+                        <icon>1d_lidar.svg</icon>
+                        <desc></desc>
+                </function>
+        </functions>
+</description>
